move filters back to respective function
